Upgrade dependencies and overhaul project configuration structure
