+++ .editorconfig
@@ -9,5 +9,14 @@ max_line_length = 120
 trim_trailing_whitespace = true
 insert_final_newline = true
 
-[*.gradle.kts]
-indent_size = 3
+end_of_line = lf
+
+[*.bat]
+end_of_line = crlf
+
+[*.md]
+indent_size = 2
+
+[*.kt]
+ij_kotlin_name_count_to_use_star_import = 99
+ij_kotlin_name_count_to_use_star_import_for_members = 99
