+++ buildSrc/build.gradle.kts
@@ -1,9 +1,3 @@
-import org.gradle.kotlin.dsl.`kotlin-dsl`
-
-repositories {
-   jcenter()
-}
-
 plugins {
    `kotlin-dsl`
 }
