+++ buildSrc/settings.gradle.kts
@@ -1,0 +1,21 @@
+rootProject.name = "buildSrc"
+
+dependencyResolutionManagement {
+   repositories {
+      mavenLocal()
+      mavenCentral()
+      maven {
+         url = uri("https://oss.sonatype.org/content/repositories/snapshots")
+      }
+   }
+}
+
+pluginManagement {
+   repositories {
+      gradlePluginPortal()
+      mavenCentral()
+      maven {
+         url = uri("https://oss.sonatype.org/content/repositories/snapshots")
+      }
+   }
+}
