+++ gradle/libs.versions.toml
@@ -1,0 +1,22 @@
+[versions]
+
+kotlin = "1.6.21"
+kotest = "5.5.5"
+embedded-kafka = "3.3.2"
+slf4j = "1.7.36"
+
+[libraries]
+
+kotlin-stdlib = { module = "org.jetbrains.kotlin:kotlin-stdlib", version.ref = "kotlin" }
+
+kotest-runner-junit5 = { module = "io.kotest:kotest-runner-junit5-jvm", version.ref = "kotest" }
+kotest-framework-api = { module = "io.kotest:kotest-framework-api", version.ref = "kotest" }
+kotest-assertions = { module = "io.kotest:kotest-assertions-core-jvm", version.ref = "kotest" }
+
+embedded-kafka = { module = "io.github.embeddedkafka:embedded-kafka_2.13", version.ref = "embedded-kafka" }
+
+slf4j-simple = { module = "org.slf4j:slf4j-simple", version.ref = "slf4j" }
+
+[plugins]
+
+kotlin-jvm = { id = "org.jetbrains.kotlin.jvm", version.ref = "kotlin" }
